Order Details Screen

+++ app/src/main/java/com/bron24/bron24_android/screens/menu_pages/orders_page/OrdersPageContract.kt
@@ -31,6 +31,6 @@ interface OrdersPageContract {
     }
 
     interface Direction{
-        suspend fun moveToInfo(orderId: Long)
+        suspend fun moveToInfo(orderId: Int)
     }
 }
